+++ src/IDealPaymentMethod.php
@@ -20,7 +20,7 @@ use Pronamic\WordPress\Pay\Plugin;
  * Company: Pronamic
  *
  * @author  Remco Tolsma
- * @version 2.1.0
+ * @version 2.1.1
  * @since   1.1.3
  */
 class IDealPaymentMethod extends PaymentMethod {
@@ -39,7 +39,7 @@ class IDealPaymentMethod extends PaymentMethod {
 	 * @param EE_Payment_Method $pm_instance Event Espresso payment method instance.
 	 */
 	public function __construct( $pm_instance = null ) {
-		$this->_gateway            = new IDealGateway46();
+		$this->_gateway            = new IDealGateway();
 		$this->_pretty_name        = __( 'iDEAL', 'pronamic_ideal' );
 		$this->_default_button_url = plugins_url( 'images/ideal/icon-64x48.png', Plugin::$file );